backend/ml_models/relighting/requirements.txt
#essentials
pip
setuptools
wheel

#cv-vision related
opencv-python
matplotlib
torch
torchvision

#TODO : remove
# SAM2
# git+https://github.com/facebookresearch/segment-anything-2.git

#transformers
transformers
diffusers
accelerate
# protobuf==3.20.3

#image procesing
OpenEXR
Imath
kornia

# Runtime helpers
requests
safetensors
numba
imageio
pillow
